We don't need to stinkin' singletons
Using the singleton pattern is effective, but because we're going to use
the `Command#requires()` method to include subsystems from our commands,
we don't need the added complexity. Furthermore, by placing the
declarations for subsystems directly into the Robot class, we can reach
these objects and don't have to obtain an instance using a class method.